chore: reorganize `data_model` crates
