+++ data_model/samples/executor_custom_data_model/src/lib.rs
@@ -5,7 +5,6 @@
 extern crate alloc;
 
 pub mod complex_isi;
-pub mod mint_rose_args;
 pub mod parameters;
 pub mod permissions;
 pub mod simple_isi;
