+++ data_model/samples/mint_rose_trigger_data_model/src/lib.rs
@@ -1,4 +1,5 @@
 //! Arguments to mint rose with args trigger
+#![no_std]
 
 use iroha_data_model::prelude::Json;
 use serde::{Deserialize, Serialize};
