--- wasm/samples/mint_rose_trigger_args/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "mint_rose_args"
-
-edition.workspace = true
-version.workspace = true
-authors.workspace = true
-
-license.workspace = true
-
-[lib]
-crate-type = ['cdylib']
-
-[dependencies]
-iroha_trigger.workspace = true
-executor_custom_data_model.workspace = true
-
-panic-halt.workspace = true
-dlmalloc.workspace = true
-getrandom.workspace = true
-
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true, default-features = false }
